+++ src/Nelmio/Alice/Loader/NativeLoader.php
@@ -12,10 +12,18 @@
 namespace Nelmio\Alice\Loader;
 
 
+use Nelmio\Alice\FileLocator\DefaultFileLocator;
+use Nelmio\Alice\FixtureBuilder\Parser\Chainable\PhpParser;
+use Nelmio\Alice\FixtureBuilder\Parser\Chainable\YamlParser;
+use Nelmio\Alice\FixtureBuilder\Parser\IncludeProcessor\DefaultIncludeProcessor;
+use Nelmio\Alice\FixtureBuilder\Parser\ParserRegistry;
+use Nelmio\Alice\FixtureBuilder\Parser\RuntimeCacheParser;
+use Nelmio\Alice\FixtureBuilder\ParserInterface;
 use Nelmio\Alice\FixtureBuilderInterface;
 use Nelmio\Alice\GeneratorInterface;
 use Nelmio\Alice\LoaderInterface;
 use Nelmio\Alice\ObjectSet;
+use Symfony\Component\Yaml\Parser as SymfonyYamlParser;
 
 /**
  * Loader implementation made to be usable without any dependency injection for quick and easy usage. For more advanced
@@ -47,9 +55,19 @@ final class NativeLoader implements LoaderInterface
 
         return $this->generator->generate($fixtureSet);
     }
-    
+
     public function __clone()
     {
         //TODO
+    }
+
+    private function getBuiltInParser(): ParserInterface
+    {
+        $parserRegistry = new ParserRegistry([
+            new YamlParser(new SymfonyYamlParser()),
+            new PhpParser(),
+        ]);
+
+        return new RuntimeCacheParser($parserRegistry, new DefaultIncludeProcessor(new DefaultFileLocator()));
     }
 }
